% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spiges_get_df.R
\name{spiges_get_df}
\alias{spiges_get_df}
\title{Get SpiGes node as dataframe}
\usage{
spiges_get_df(x, node, variables = NULL)
}
\arguments{
\item{x}{A string, a connection, or a raw vector. See \code{xml2::read_xml}}

\item{node}{Name of the node as string, see \code{get_node_names}}

\item{variables}{Name of the variables, by default returns all variables}
}
\value{
data.frame/tibble
}
\description{
Return a tibble/dataframe of a SpiGes node name. By default, all variables
are returned. Variables can be selected using the `variables` argument.
}
