% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spiges_get_name_nodes.R
\name{spiges_get_name_nodes}
\alias{spiges_get_name_nodes}
\title{Get node names available for data extraction of a SpiGes XML}
\usage{
spiges_get_name_nodes(x)
}
\arguments{
\item{x}{A string, a connection, or a raw vector. See \code{xml2::read_xml}}
}
\value{
vector of node names
}
\description{
Return main and children node names of an SpiGes XML. The node names
"Unternehmen", "Standort", "Fall", "Kantonsdaten" are removed as the function
spiges_get_df() doesn't provide support to extract their related data.
}
