% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spiges_get_name_variables.R
\name{spiges_get_name_variables}
\alias{spiges_get_name_variables}
\title{List of official variable names by node}
\usage{
spiges_get_name_variables()
}
\value{
a list of variable names
}
\description{
Return a list containing the list of the variables names for each SpiGes XML.
The names of the list are equivalent of the node names returned by
}
\details{
Source file: \url{https://www.bfs.admin.ch/bfs/de/home/statistiken/gesundheit/gesundheitswesen/projekt-spiges.assetdetail.28127962.html}
}
